Reorganise demo pages
