--- packages/scxa-faceted-search-results/html/EpisodesHeader.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const EpisodesHeader = () =>
-  <div style={{textAlign: `center`}}>
-    <b>Episode Title</b>
-  </div>
-
-export default EpisodesHeader
