+++ packages/scxa-faceted-search-results/html/fetch-loader/demo.js
@@ -1,17 +1,22 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import FetchLoader from '../../src/FetchLoader'
+import { withFetchLoader } from '@ebi-gene-expression-group/atlas-react-fetch-loader'
+import FacetedSearchContainer from '../../src/index'
+
 import ExperimentCard from './ExperimentCard'
 import ExperimentsHeader from './ExperimentsHeader'
 
+const FacetedSearchContainerWithFetchLoader = withFetchLoader(FacetedSearchContainer)
+
 const render = (options, target) => {
   ReactDOM.render(
-    <FetchLoader
+    <FacetedSearchContainerWithFetchLoader
       {...options}
       sortTitle={`markerGenes`}
       ResultElementClass={ExperimentCard}
-      ResultsHeaderClass={ExperimentsHeader} />,
+      ResultsHeaderClass={ExperimentsHeader}
+    />,
     document.getElementById(target))
 }
