+++ packages/scxa-faceted-search-results/html/fetch-loader/error.html
@@ -16,12 +16,12 @@
 </div>
 </body>
 
-<script src="../../dist/vendors.bundle.js"></script>
-<script src="../../dist/fetchLoaderDemo.bundle.js"></script>
+<script src="/dist/vendors.bundle.js"></script>
+<script src="/dist/facetedSearchContainerWithFetchLoaderDemo.bundle.js"></script>
 
-<!-- Set to http://localhost:8080/gxa/ or http://localhost:8080/gxa/sc/ -- Remember the trailing slash! -->
 <script>
-    fetchLoaderDemo.render({
+    facetedSearchContainerWithFetchLoaderDemo.render({
+      // Set to http://localhost:8080/gxa/, http://localhost:8080/gxa/sc/ or wherever, but remember the trailing slash!
       host: 'https://www.ebi.ac.uk/gxa/',
       resource: 'foobar'
       // This is an example in which the component will silently fail, since it receives a valid JSON response but it
