+++ packages/scxa-faceted-search-results/html/index.html
@@ -15,11 +15,11 @@
       </li>
 
       <li>
-        <a href="fetch-loader/demo.html">FetchLoader (retrieves real data from <code>localhost:8080/gxa/sc</code>)</a>
+        <a href="fetch-loader">FetchLoader (retrieves real data from <code>localhost:8080/gxa/sc</code>)</a>
       </li>
 
       <li>
-        <a href="faceted-search-container-demo.html">FacetedSearchContainer (pass <code>props</code> directly)</a>
+        <a href="container">FacetedSearchContainer (pass <code>props</code> directly, no fetch)</a>
       </li>
     </ul>
   </div>
